Simplify snippet optional parameter navigation
When the tab key is pressed, the next part of the
snipper is highlighted for easy editing. This
change makes it so that first the whole optional
syntax element is selected, including any
surrounding separator, line break, etc., so that
it may be deleted easily if wanted, and should
the user want to use the optional parameter, an
extra tab press will select only the value,
preserving the aforementioned separators.